requirements.txt
APScheduler==3.11.0
asgiref==3.9.2
beautifulsoup4==4.14.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.10
jwt==1.4.0
pillow==11.3.0
pycparser==2.23
python-dotenv==1.1.1
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
